Add LanguageViewModelService in the admin area ( #390)

+++ Grand.Web/Areas/Admin/Controllers/LanguageController.cs
@@ -1,21 +1,18 @@
-﻿using Grand.Core;
-using Grand.Core.Domain.Localization;
+﻿using Grand.Core.Domain.Localization;
 using Grand.Framework.Extensions;
 using Grand.Framework.Kendoui;
 using Grand.Framework.Mvc;
 using Grand.Framework.Mvc.Filters;
 using Grand.Framework.Security;
-using Grand.Services.Directory;
 using Grand.Services.Localization;
 using Grand.Services.Security;
 using Grand.Services.Stores;
 using Grand.Web.Areas.Admin.Extensions;
 using Grand.Web.Areas.Admin.Models.Localization;
+using Grand.Web.Areas.Admin.Services;
 using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
-using Microsoft.AspNetCore.Mvc.Rendering;
 using System;
-using System.Collections.Generic;
 using System.IO;
 using System.Linq;
 using System.Text;
@@ -25,76 +22,32 @@ namespace Grand.Web.Areas.Admin.Controllers
     public partial class LanguageController : BaseAdminController
     {
         #region Fields
-
+        private readonly ILanguageViewModelService _languageViewModelService;
         private readonly ILanguageService _languageService;
         private readonly ILocalizationService _localizationService;
-        private readonly ICurrencyService _currencyService;
         private readonly IStoreService _storeService;
-        private readonly IStoreMappingService _storeMappingService;
         private readonly IPermissionService _permissionService;
-        private readonly IWebHelper _webHelper;
 
         #endregion
 
         #region Constructors
 
-        public LanguageController(ILanguageService languageService,
+        public LanguageController(
+            ILanguageViewModelService languageViewModelService,
+            ILanguageService languageService,
             ILocalizationService localizationService,
-            ICurrencyService currencyService,
             IStoreService storeService,
-            IStoreMappingService storeMappingService,
-            IPermissionService permissionService,
-            IWebHelper webHelper)
+            IPermissionService permissionService)
         {
+            this._languageViewModelService = languageViewModelService;
             this._localizationService = localizationService;
             this._languageService = languageService;
-            this._currencyService = currencyService;
             this._storeService = storeService;
-            this._storeMappingService = storeMappingService;
             this._permissionService = permissionService;
-            this._webHelper = webHelper;
         }
 
         #endregion
 
-        #region Utilities
-
-        [NonAction]
-        protected virtual void PrepareFlagsModel(LanguageModel model)
-        {
-            if (model == null)
-                throw new ArgumentNullException("model");
-
-            model.FlagFileNames = Directory
-                .EnumerateFiles(CommonHelper.MapPath("~/wwwroot/content/images/flags/"), "*.png", SearchOption.TopDirectoryOnly)
-                .Select(Path.GetFileName)
-                .ToList();
-        }
-
-        [NonAction]
-        protected virtual void PrepareCurrenciesModel(LanguageModel model)
-        {
-            if (model == null)
-                throw new ArgumentNullException("model");
-
-            //templates
-            model.AvailableCurrencies.Add(new SelectListItem
-            {
-                Text = "---",
-                Value = ""
-            });
-            var currencies = _currencyService.GetAllCurrencies(true);
-            foreach (var currency in currencies)
-            {
-                model.AvailableCurrencies.Add(new SelectListItem
-                {
-                    Text = currency.Name,
-                    Value = currency.Id.ToString()
-                });
-            }
-        }
-
-        #endregion
 
         #region Languages
 
@@ -135,9 +88,9 @@ namespace Grand.Web.Areas.Admin.Controllers
             //Stores
             model.PrepareStoresMappingModel(null, false, _storeService);
             //currencies
-            PrepareCurrenciesModel(model);
+            _languageViewModelService.PrepareCurrenciesModel(model);
             //flags
-            PrepareFlagsModel(model);
+            _languageViewModelService.PrepareFlagsModel(model);
             //default values
             model.Published = true;
             return View(model);
@@ -151,22 +104,18 @@ namespace Grand.Web.Areas.Admin.Controllers
 
             if (ModelState.IsValid)
             {
-                var language = model.ToEntity();
-                language.Stores = model.SelectedStoreIds != null ? model.SelectedStoreIds.ToList() : new List<string>();
-                _languageService.InsertLanguage(language);
-
+                var language = _languageViewModelService.InsertLanguageModel(model);
                 SuccessNotification(_localizationService.GetResource("Admin.Configuration.Languages.Added"));
                 return continueEditing ? RedirectToAction("Edit", new { id = language.Id }) : RedirectToAction("List");
             }
 
             //If we got this far, something failed, redisplay form
-
             //Stores
             model.PrepareStoresMappingModel(null, true, _storeService);
             //currencies
-            PrepareCurrenciesModel(model);
+            _languageViewModelService.PrepareCurrenciesModel(model);
             //flags
-            PrepareFlagsModel(model);
+            _languageViewModelService.PrepareFlagsModel(model);
 
             return View(model);
         }
@@ -185,9 +134,9 @@ namespace Grand.Web.Areas.Admin.Controllers
             //Stores
             model.PrepareStoresMappingModel(language, false, _storeService);
             //currencies
-            PrepareCurrenciesModel(model);
+            _languageViewModelService.PrepareCurrenciesModel(model);
             //flags
-            PrepareFlagsModel(model);
+            _languageViewModelService.PrepareFlagsModel(model);
 
             return View(model);
         }
@@ -214,12 +163,7 @@ namespace Grand.Web.Areas.Admin.Controllers
                     return RedirectToAction("Edit", new { id = language.Id });
                 }
 
-                //update
-                language = model.ToEntity(language);
-                language.Stores = model.SelectedStoreIds != null ? model.SelectedStoreIds.ToList() : new List<string>();
-
-                _languageService.UpdateLanguage(language);
-
+                language = _languageViewModelService.UpdateLanguageModel(language, model);
                 //notification
                 SuccessNotification(_localizationService.GetResource("Admin.Configuration.Languages.Updated"));
                 if (continueEditing)
@@ -231,15 +175,13 @@ namespace Grand.Web.Areas.Admin.Controllers
                 }
                 return RedirectToAction("List");
             }
-
             //If we got this far, something failed, redisplay form
-
             //Stores
             model.PrepareStoresMappingModel(language, true, _storeService);
             //currencies
-            PrepareCurrenciesModel(model);
+            _languageViewModelService.PrepareCurrenciesModel(model);
             //flags
-            PrepareFlagsModel(model);
+            _languageViewModelService.PrepareFlagsModel(model);
 
             return View(model);
         }
@@ -264,11 +206,16 @@ namespace Grand.Web.Areas.Admin.Controllers
             }
 
             //delete
-            _languageService.DeleteLanguage(language);
+            if (ModelState.IsValid)
+            {
+                _languageService.DeleteLanguage(language);
 
-            //notification
-            SuccessNotification(_localizationService.GetResource("Admin.Configuration.Languages.Deleted"));
-            return RedirectToAction("List");
+                //notification
+                SuccessNotification(_localizationService.GetResource("Admin.Configuration.Languages.Deleted"));
+                return RedirectToAction("List");
+            }
+            ErrorNotification(ModelState);
+            return RedirectToAction("Edit", new { id = language.Id });
         }
 
         #endregion
@@ -283,32 +230,11 @@ namespace Grand.Web.Areas.Admin.Controllers
             if (!_permissionService.Authorize(StandardPermissionProvider.ManageLanguages))
                 return AccessDeniedView();
 
-            var language = _languageService.GetLanguageById(languageId);
-
-            var resources = _localizationService
-                .GetAllResources(languageId)
-                .OrderBy(x => x.ResourceName)
-                .Select(x => new LanguageResourceModel
-                {
-                    LanguageId = languageId,
-                    Id = x.Id,
-                    Name = x.ResourceName,
-                    Value = x.ResourceValue,
-                });
-
-            if (model != null)
-            {
-                if (!string.IsNullOrEmpty(model.ResourceName))
-                    resources = resources.Where(x => x.Name.ToLowerInvariant().Contains(model.ResourceName.ToLowerInvariant()));
-                if (!string.IsNullOrEmpty(model.ResourceValue))
-                    resources = resources.Where(x => x.Value.ToLowerInvariant().Contains(model.ResourceValue.ToLowerInvariant()));
-            }
-            resources = resources.AsQueryable();
-
+            var resources = _languageViewModelService.PrepareLanguageResourceModel(model, languageId, command.Page, command.PageSize);
             var gridModel = new DataSourceResult
             {
-                Data = resources.PagedForCommand(command),
-                Total = resources.Count()
+                Data = resources.languageResourceModels.ToList(),
+                Total = resources.totalCount
             };
 
             return Json(gridModel);
@@ -320,63 +246,32 @@ namespace Grand.Web.Areas.Admin.Controllers
             if (!_permissionService.Authorize(StandardPermissionProvider.ManageLanguages))
                 return AccessDeniedView();
 
-            if (model.Name != null)
-                model.Name = model.Name.Trim();
-            if (model.Value != null)
-                model.Value = model.Value.Trim();
-
             if (!ModelState.IsValid)
             {
                 return Json(new DataSourceResult { Errors = ModelState.SerializeErrors() });
             }
 
-            var resource = _localizationService.GetLocaleStringResourceById(model.Id);
-            // if the resourceName changed, ensure it isn't being used by another resource
-            if (!resource.ResourceName.Equals(model.Name, StringComparison.OrdinalIgnoreCase))
-            {
-                var res = _localizationService.GetLocaleStringResourceByName(model.Name, model.LanguageId, false);
-                if (res != null && res.Id != resource.Id)
-                {
-                    return Json(new DataSourceResult { Errors = string.Format(_localizationService.GetResource("Admin.Configuration.Languages.Resources.NameAlreadyExists"), res.ResourceName) });
-                }
-            }
-
-            resource.ResourceName = model.Name;
-            resource.ResourceValue = model.Value;
-            _localizationService.UpdateLocaleStringResource(resource);
-
+            var result = _languageViewModelService.UpdateLanguageResourceModel(model);
+            if (result.error)
+                return ErrorForKendoGridJson(result.message);
             return new NullJsonResult();
         }
 
         [HttpPost]
-        public IActionResult ResourceAdd(string languageId, LanguageResourceModel model)
+        public IActionResult ResourceAdd(LanguageResourceModel model)
         {
             if (!_permissionService.Authorize(StandardPermissionProvider.ManageLanguages))
                 return AccessDeniedView();
-
-            if (model.Name != null)
-                model.Name = model.Name.Trim();
-            if (model.Value != null)
-                model.Value = model.Value.Trim();
 
             if (!ModelState.IsValid)
             {
                 return Json(new DataSourceResult { Errors = ModelState.SerializeErrors() });
             }
-
-            var res = _localizationService.GetLocaleStringResourceByName(model.Name, model.LanguageId, false);
-            if (res == null)
+            var result = _languageViewModelService.InsertLanguageResourceModel(model);
+            if (result.error)
             {
-                var resource = new LocaleStringResource { LanguageId = languageId };
-                resource.ResourceName = model.Name;
-                resource.ResourceValue = model.Value;
-                _localizationService.InsertLocaleStringResource(resource);
+                return ErrorForKendoGridJson(result.message);
             }
-            else
-            {
-                return Json(new DataSourceResult { Errors = string.Format(_localizationService.GetResource("Admin.Configuration.Languages.Resources.NameAlreadyExists"), model.Name) });
-            }
-
             return new NullJsonResult();
         }
 
@@ -389,9 +284,12 @@ namespace Grand.Web.Areas.Admin.Controllers
             var resource = _localizationService.GetLocaleStringResourceById(id);
             if (resource == null)
                 throw new ArgumentException("No resource found with the specified id");
-            _localizationService.DeleteLocaleStringResource(resource);
-
-            return new NullJsonResult();
+            if (ModelState.IsValid)
+            {
+                _localizationService.DeleteLocaleStringResource(resource);
+                return new NullJsonResult();
+            }
+            return ErrorForKendoGridJson(ModelState);
         }
 
         #endregion
